Shuffle dependencies around to add modified nak
Well, this is a hack. I built a modified version of nak-1.1.0.jar
to add undo methods to all of the PointTransformer classes, and then I
added the modified nak to the lib directory. Then I also checked in
chalk and all of its dependencies to prevent it from being downloaded
and pulling along the unmodified nak with it.